SPHS update
Now serves html files in the browser instead of as a download. Index argument will accept local directory files and full file paths now. Including a basic index page with a shutdown button. Adding a log displayer to the code at some point. Will likely set the content of index.html to the default GET response in the future.